Squashed commit of the following:
commit ca708a96839467befcab57db060d641e4ade4ac9
Author: Miku
Date:   Sun Nov 10 13:19:04 2019 +1200

    Update dictionaries, allow to log in.

commit 064eb25199fcfa21508439625f144d45bde0c91c
Author: Miku
Date:   Sat Nov 9 16:23:38 2019 +1200

    Update managers.

commit 03a1733df47314f12e25822696bfaad5b1f09a3d
Author: Miku
Date:   Sat Nov 9 13:40:29 2019 +1200

    Replace MySql calls with EF.

commit 9fa82966cd346d1c8956077bdfbe66f4e209de76
Author: Miku
Date:   Sat Nov 9 01:03:03 2019 +1200

    Update language version. Update nuget packages. Migrate core entities to EF.

commit 3113f82855087f9547edd22cefd1a874f6625dfd
Author: Miku
Date:   Fri Nov 8 01:11:23 2019 +1200

    Migrate login server to ef. Add DB entities.

commit 626f09c03858bfe2f4b431ae3788e1aea701a607
Author: Miku
Date:   Wed Nov 6 22:37:17 2019 +1200

    Refactor configuration files to use connection strings.

commit 78e150f28a289194df33bdefe0da157f37145c0c
Author: Miku
Date:   Wed Nov 6 21:56:46 2019 +1200

    Add update script file orders numbers. Support MySQL Server 5.7.

commit be0f3b81efd95798aa8949809eb6d7f1345e7a54
Author: Miku
Date:   Wed Nov 6 20:59:15 2019 +1200

    Migrate to .net core 3.0

+++ AAEmu.Game/Core/Managers/SkillManager.cs
@@ -161,7 +161,7 @@ namespace AAEmu.Game.Core.Managers
                             template.AbilityId = reader.GetByte("ability_id");
                             template.ManaCost = reader.GetInt32("mana_cost");
                             template.TimingId = reader.GetInt32("timing_id");
-                            template.CooldownTime = reader.GetInt32("cooldown_time");
+                            template.CooldownTime = reader.GetUInt64("cooldown_time");
                             template.CastingTime = reader.GetInt32("casting_time");
                             template.IgnoreGlobalCooldown = reader.GetBoolean("ignore_global_cooldown", true);
                             template.EffectDelay = reader.GetInt32("effect_delay");
@@ -325,7 +325,7 @@ namespace AAEmu.Game.Core.Managers
                             template.Id = reader.GetUInt32("id");
                             template.AnimStartId = reader.GetUInt32("anim_start_id", 0);
                             template.AnimEndId = reader.GetUInt32("anim_end_id", 0);
-                            template.Duration = reader.GetInt32("duration");
+                            template.Duration = reader.GetInt64("duration");
                             template.Tick = reader.GetDouble("tick");
                             template.Silence = reader.GetBoolean("silence", true);
                             template.Root = reader.GetBoolean("root", true);
@@ -357,7 +357,7 @@ namespace AAEmu.Game.Core.Managers
                             template.SpellImmune = reader.GetBoolean("spell_immune", true);
                             template.RangedImmune = reader.GetBoolean("ranged_immune", true);
                             template.SiegeImmune = reader.GetBoolean("siege_immune", true);
-                            template.ImmuneDamage = reader.GetInt32("immune_damage");
+                            template.ImmuneDamage = reader.GetInt64("immune_damage");
                             var value = reader.GetString("skill_controller_id", "0");
                             template.SkillControllerId = value.Contains("null") ? 0 : uint.Parse(value);
                             template.ResurrectionHealth = reader.GetInt32("resurrection_health");
@@ -1158,10 +1158,10 @@ namespace AAEmu.Game.Core.Managers
                             var template = new SpecialEffect();
                             template.Id = reader.GetUInt32("id");
                             template.SpecialEffectTypeId = (SpecialEffectType)reader.GetInt32("special_effect_type_id");
-                            template.Value1 = reader.GetInt32("value1");
-                            template.Value2 = reader.GetInt32("value2");
-                            template.Value3 = reader.GetInt32("value3");
-                            template.Value4 = reader.GetInt32("value4");
+                            template.Value1 = reader.GetInt64("value1");
+                            template.Value2 = reader.GetInt64("value2");
+                            template.Value3 = reader.GetInt64("value3");
+                            template.Value4 = reader.GetInt64("value4");
                             _effects["SpecialEffect"].Add(template.Id, template);
                         }
                     }
