+++ AAEmu.Game/Models/Game/Skills/Effects/SpecialEffect.cs
@@ -13,7 +13,7 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
     public interface ISpecialEffect
     {
         void Execute(Unit caster, SkillCaster casterObj, BaseUnit target, SkillCastTarget targetObj, CastAction castObj,
-            Skill skill, SkillObject skillObject, DateTime time, int Value1, int Value2, int Value3, int Value4);
+            Skill skill, SkillObject skillObject, DateTime time, long Value1, long Value2, long Value3, long Value4);
     }
 
     public enum SpecialEffectType
@@ -140,10 +140,10 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
     public class SpecialEffect : EffectTemplate
     {
         public SpecialEffectType SpecialEffectTypeId { get; set; } // TODO inspect
-        public int Value1 { get; set; }
-        public int Value2 { get; set; }
-        public int Value3 { get; set; }
-        public int Value4 { get; set; }
+        public long Value1 { get; set; }
+        public long Value2 { get; set; }
+        public long Value3 { get; set; }
+        public long Value4 { get; set; }
 
         public override bool OnActionTime => false;
